🛡️ Sentinel: [CRITICAL] Fix API Key Exposure
🚨 Severity: CRITICAL
💡 Vulnerability: GITHUB_TOKEN and JULES_API_KEY were partially validated and redacted across different fleet scripts, leading to potential sensitive key exposure in error logs and missing validations where they were needed.
🎯 Impact: Secrets exposure in logs can lead to full repository compromise or unauthorized API usage.
🔧 Fix: Centralized token validation and redaction logic in `scripts/fleet/env.ts`. Overrode `console.log`, `console.error`, and `console.warn` across all fleet entry points to automatically redact tokens using safe stream writing instead of mutating objects.
✅ Verification: Ran `bun test` and `bun build --target=node` locally. Reviewed `env.ts` redaction wrapper to ensure format strings don't bypass checks.
